Move iterators implementation.
