+++ ksp_plugin/interface_renderer.cpp
@@ -3,13 +3,17 @@
 
 #include "journal/method.hpp"
 #include "journal/profiles.hpp"
+#include "ksp_plugin/iterators.hpp"
 #include "ksp_plugin/plugin.hpp"
 #include "ksp_plugin/renderer.hpp"
+#include "physics/discrete_trajectory.hpp"
 
 namespace principia {
 namespace interface {
 
 using ksp_plugin::Renderer;
+using ksp_plugin::TypedIterator;
+using physics::DiscreteTrajectory;
 
 namespace {
