+++ ksp_plugin/ksp_plugin.vcxproj.filters
@@ -68,6 +68,9 @@
     <ClInclude Include="iterators.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="iterators_body.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="interface.cpp">
